==> notebooks/01-mas-eda-wrangling.ipynb <==
# **Data Wrangling**

## **Imports**

import os
import numpy as np
import pandas as pd

# from sklearn.experimental import enable_iterative_imputer
# from sklearn.impute import IterativeImputer

## **Paths & Global Variables**

ROOT_PATH = r"C:\Users\mario\OneDrive\Documents\Work\Side Hustles\Kaggle\titanic"

import sys

sys.path.append(ROOT_PATH)

from titanic.config import RAW_DATA_DIR, INTERIM_DATA_DIR

## **Reading Data**

os.chdir(RAW_DATA_DIR)
train_df = pd.read_csv("train.csv")
test_df = pd.read_csv("test.csv")

train_df["split"] = "train"
test_df["split"] = "test"
test_df["Survived"] = 0

raw_df = pd.concat([train_df, test_df], axis=0)

## **Data Diagnosis**

raw_df.head()

raw_df.shape

raw_df.columns

## **Column Name Consistency**

raw_df.columns = map(str.lower, raw_df.columns)

# axis=1` refers to the columns, `axis=0` would refer to the rows
column_new_names_dict = {
    "passengerid": "passenger_id",
    "pclass": "passenger_class",
    "name": "full_name",
    "sibsp": "siblings_spouses",
    "parch": "parents_children",
}

raw_df = raw_df.rename(column_new_names_dict, axis=1)

raw_df.columns

## **Column Profiling**

raw_df.dtypes

raw_df.nunique()

raw_df.isna().sum()

raw_df.info()

## **Data Types**

### **Numerical Variables**

#### **Discrete Variables**

print(raw_df.passenger_id.nunique())
print(raw_df.passenger_id.unique())

print(raw_df.age.nunique())
print(raw_df.age.unique())

print(raw_df.siblings_spouses.nunique())
print(raw_df.siblings_spouses.unique())

print(raw_df.parents_children.nunique())
print(raw_df.parents_children.unique())

#### **Continuous Variables**

print(raw_df.fare.nunique())
print(raw_df.fare.unique())

### **Categorical Variables**

#### **Binary Variables**

print(raw_df.survived.nunique())
print(raw_df.survived.unique())

#### **Ordinal Variables**

print(raw_df.passenger_class.nunique())
print(raw_df.passenger_class.unique())

#### **Nominal Variables**

print(raw_df.sex.nunique())
print(raw_df.sex.unique())

print(raw_df.embarked.nunique())
print(raw_df.embarked.unique())

print(raw_df.cabin.nunique())
print(raw_df.cabin.unique())

print(raw_df.full_name.nunique())
print(raw_df.full_name.unique())

print(raw_df.ticket.nunique())
print(raw_df.ticket.unique())

## **Removing Duplicates**

INDEX_COLS_LIST = ["passenger_id"]

duplicates = raw_df[raw_df[INDEX_COLS_LIST].duplicated()]
duplicates

print(raw_df.shape)
raw_df.drop_duplicates(subset=INDEX_COLS_LIST, keep="first", inplace=True)
print(raw_df.shape)

## **Indexing**


raw_df.set_index(INDEX_COLS_LIST, inplace=True)
raw_df.head(1)

## **Missing Values**

### **Deletion**

#### **Listwise Deletion**

# raw_df.dropna(inplace=True)

#### **Pairwise Deletion**

raw_df.dropna(subset=["embarked"], inplace=True)

#### **Dropping Columns**

### **Single Imputation**

raw_df.fillna(value={"cabin": "Unknown"}, inplace=True)

raw_df["age"] = raw_df.groupby(["passenger_class", "sex", "embarked"])["age"].transform(
    lambda x: x.fillna(x.mean())
)

raw_df["fare"] = raw_df.groupby(["passenger_class", "sex", "embarked"])[
    "fare"
].transform(lambda x: x.fillna(x.mean()))

### **Multiple Imputation** 🔴

## **Splitting Columns**

raw_df["surname"] = raw_df["full_name"].str.split(",").str[0]
raw_df["surname"].unique()

raw_df["title"] = raw_df["full_name"].str.split(",").str[1].str.split(".").str[0]
raw_df["title"] = raw_df["title"].str.strip()
raw_df["title"].unique()

raw_df["first_name"] = raw_df["full_name"].str.split(",").str[1].str.split(".").str[1]
raw_df["first_name"] = raw_df["first_name"].str.strip()
raw_df["first_name"].unique()

## **String Parsing** 🔴

raw_df["list_cabins"] = raw_df["cabin"].str.split(r"\s")

# Could be Feature Engineering

raw_df["num_rooms"] = raw_df["list_cabins"].apply(
    lambda x: len(x) if x[0] != "Unknown" else 0
)
raw_df["num_rooms"] = raw_df["num_rooms"].astype(int)
raw_df["num_rooms"].unique()

raw_df.drop(columns=["list_cabins"], inplace=True)

raw_df["cabin_level_1"] = raw_df["cabin"].str[0]
raw_df["cabin_level_1"].unique()

# Define a function to replace 'U' with the mode of the group
def replace_U_with_mode(group):
    mode_value = group[~group.isin(["U", "T"])].mode()
    if not mode_value.empty:
        group[group.isin(["U", "T"])] = mode_value[0]
    return group


# Apply the function to each group
raw_df["cabin_level_1"] = raw_df.groupby(["passenger_class", "embarked"])[
    "cabin_level_1"
].transform(replace_U_with_mode)
raw_df.cabin_level_1.value_counts()

raw_df["cabin_level_2"] = raw_df["cabin"].str[1:]
raw_df["cabin_level_2"] = raw_df["cabin_level_2"].str.extract(r"(\d+)")
raw_df["cabin_level_2"] = raw_df["cabin_level_2"].fillna(0)
raw_df["cabin_level_2"] = raw_df["cabin_level_2"].astype(int)
raw_df["cabin_level_2"].unique()

## **Data Type Transformations**

raw_df.info()

raw_df.columns

### **Discrete Variables**

DISCRETE_VARS_LIST = ["siblings_spouses", "parents_children", "num_rooms"]

for col in DISCRETE_VARS_LIST:
    raw_df[col] = raw_df[col].astype("int")

### **Continuous Variables**

Continuous (numerical) variables should usually be stored as the `float` data type because they allow us to store decimal values.

CONT_VARS_LIST = ["age", "fare"]

for col in CONT_VARS_LIST:
    raw_df[col] = raw_df[col].astype("float")

### **Binary Variables**

BOOL_VARS_LIST = ["survived"]
for col in BOOL_VARS_LIST:
    raw_df[col] = raw_df[col].astype("bool")

### **Ordinal Variables**

ORD_VARS_DICT = {"passenger_class": {3: "Lower", 2: "Middle", 1: "Upper"}}

for col in ORD_VARS_DICT.keys():
    raw_df[col] = raw_df[col].astype("category")
    print(raw_df[col].unique())
    print(raw_df[col].sort_values().head())

for col, ordered_dict in ORD_VARS_DICT.items():
    print(col)
    raw_df[col] = pd.Categorical(raw_df[col], ordered_dict.keys(), ordered=True)
    raw_df[col] = raw_df[col].cat.rename_categories(ordered_dict.values())
    print(raw_df[col].unique())
    print(raw_df[col].sort_values().head())

### **Nominal Variables**

NOM_VARS_LIST = [
    "full_name",
    "surname",
    "title",
    "first_name",
    "cabin",
    "ticket",
    "cabin_level_2",
]

for col in NOM_VARS_LIST:
    raw_df[col] = raw_df[col].astype("string")

NOM_CAT_VARS_LIST = [
    "sex",
    "embarked",
    "cabin_level_1",
]

for col in NOM_CAT_VARS_LIST:
    raw_df[col] = raw_df[col].astype("category")

raw_df.dtypes

## **Reorder Columns**

new_col_order = []
feature_types = ["int", "float", np.bool, "category", "string", "object"]

for feature_type in feature_types:
    new_col_order += raw_df.select_dtypes(include=[feature_type]).columns.to_list()

raw_df = raw_df[new_col_order]
raw_df.info()

os.chdir(INTERIM_DATA_DIR)
raw_df.to_csv("titanic_cleaned.csv")
raw_df.to_parquet("titanic_cleaned.parquet")
