==> notebooks/04-mas-train-test-split.ipynb <==
## **Imports**

import os
import pandas as pd

from sklearn.linear_model import LogisticRegression
from sklearn.neighbors import KNeighborsClassifier
from sklearn.tree import DecisionTreeClassifier

from sklearn.model_selection import StratifiedKFold, cross_val_score, train_test_split

## **Paths & Global Variables**

ROOT_PATH = r"C:\Users\mario\OneDrive\Documents\Work\Side Hustles\Kaggle\titanic"

import sys

sys.path.append(ROOT_PATH)

from titanic.config import PROCESSED_DATA_DIR

## **Reading Data**

os.chdir(PROCESSED_DATA_DIR)
final_df = pd.read_parquet("titanic_processed.parquet")
final_df.head(1)

final_df.info()

final_df.columns

# **Training, Testing, Validation Split**

numerical_d_features = [
    "siblings_spouses",
    "parents_children",
    "num_rooms",
]

final_df[numerical_d_features].head(1)

numerical_c_features = [
    "age",
    "fare",
]

final_df[numerical_c_features].head(1)

numerical_c_scaled_features = [
    "age_scaled",
    "age_min_max",
    "fare_log",
]

final_df[numerical_c_scaled_features].head(1)

binary_features = [
    "survived",
]

final_df[binary_features].head(1)

ordinal_features = [
    "passenger_class",
    "age_group",
]

final_df[ordinal_features].head(1)

ordinal_encoded_features = [
    "passenger_class_ord",
    "age_group_ord",
]

final_df[ordinal_encoded_features].head(1)

nominal_features = [
    "full_name",
    "ticket",
    "cabin",
    "surname",
    "title",
    "first_name",
    "cabin_level_2",
    "sex",
    "embarked",
]

final_df[nominal_features].head(1)

label_encoded_features = [
    "full_name_lcode",
    "surname_lcode",
    "title_lcode",
    "first_name_lcode",
    "cabin_lcode",
    "ticket_lcode",
    "cabin_level_2_lcode",
    # "sex_lcode",
    # "embarked_lcode",
    # "cabin_level_1_lcode",
]

final_df[label_encoded_features].head(1)

one_hot_encoded_features = [
    # "embarked_C",
    "embarked_Q",
    "embarked_S",
    "sex_male",
    # "cabin_level_1_A",
    "cabin_level_1_B",
    "cabin_level_1_C",
    "cabin_level_1_D",
    "cabin_level_1_E",
    "cabin_level_1_F",
    "cabin_level_1_G",
]

final_df[one_hot_encoded_features].head(1)

target_encoded_features = [
    "full_name_target",
    "surname_target",
    "title_target",
    "first_name_target",
    "cabin_target",
    "ticket_target",
    "cabin_level_2_target",
]

final_df[target_encoded_features].head(5)

target_variable = "survived"
selected_features = (
    numerical_d_features
    # + numerical_c_features
    + numerical_c_scaled_features
    # + binary_features
    # + ordinal_features
    + ordinal_encoded_features
    # + nominal_features
    # + label_encoded_features
    + one_hot_encoded_features
    # + target_encoded_features
)

y = final_df[final_df["split"] == "train"][target_variable]
X = final_df[final_df["split"] == "train"][selected_features]

y_test = final_df[final_df["split"] == "test"][target_variable]
X_test = final_df[final_df["split"] == "test"][selected_features]

X_train, X_val, y_train, y_val = train_test_split(
    X, y, test_size=0.2, random_state=42
)

print(len(X_train), len(y_train))
print(len(X_val), len(y_val))
print(len(X_test), len(y_test))

## **Positivity Rate**

print("Train positivity rate: ")
print(sum(y_train) / y_train.shape[0])
print("Test positivity rate: ")
print(sum(y_val) / y_val.shape[0])

## **Stratify Sampling**

# Stratify on y due to class imbalance
X_train, X_val, y_train, y_val = train_test_split(
    X, y, test_size=0.2, random_state=42, stratify=y
)

print("Stratified train positivity rate: ")
print(sum(y_train) / y_train.shape[0])

print("Stratified test positivity rate: ")
print(sum(y_val) / y_val.shape[0])

os.chdir(PROCESSED_DATA_DIR)

X_train.to_parquet("X_train.parquet")
X_val.to_parquet("X_val.parquet")
X_test.to_parquet("X_test.parquet")

y_train.to_frame().to_parquet("y_train.parquet")
y_val.to_frame().to_parquet("y_val.parquet")
y_test.to_frame().to_parquet("y_test.parquet")

## **Stratified K-Fold**

skf = StratifiedKFold(n_splits=5, shuffle=True, random_state=42)

log_reg = LogisticRegression(class_weight="balanced")
knn = KNeighborsClassifier(n_neighbors=15)
dtree = DecisionTreeClassifier(max_depth=8, ccp_alpha=0.01,criterion='gini')

models = [log_reg, knn, dtree]

for model in models:
    scores = cross_val_score(model, X, y, cv=skf, scoring="roc_auc")
    print(f"Model: {model}")
    print(f"Stratified CV scores: {scores}")
    print(f"Mean AUC: {scores.mean():.4f}\n")
