==> notebooks/08-mas-model-testing.ipynb <==
## **Imports**

import os
import joblib
import numpy as np
import pandas as pd

from sklearn.metrics import confusion_matrix, accuracy_score, precision_score, recall_score, f1_score

## **Paths & Global Variables**

ROOT_PATH = r"C:\Users\mario\OneDrive\Documents\Work\Side Hustles\Kaggle\titanic"

import sys

sys.path.append(ROOT_PATH)

from titanic.config import PROCESSED_DATA_DIR, MODELS_DIR

## **Reading Data**

os.chdir(PROCESSED_DATA_DIR)

X_test = pd.read_parquet("X_test.parquet")
y_test = pd.read_parquet("y_test.parquet")

os.chdir(MODELS_DIR)
logreg_model = joblib.load("log_reg-model.joblib")
knn_model = joblib.load("knn-model.joblib")
dtree_model = joblib.load("dtree-model.joblib")

## **Model Validation**

cols_to_remove = ["parents_children","num_rooms","age_min_max","age_group_ord","passenger_class_ord"]
X_test_temp = X_test.copy()
X_test_temp.drop(columns=cols_to_remove, inplace=True)

y_predicted = logreg_model.predict(X_test_temp)

## **Evaluation Metrics**

test_conf_matrix = pd.DataFrame(
    confusion_matrix(y_test, y_predicted),
    index=["actual no", "actual yes"],
    columns=["predicted no", "predicted yes"],
)

test_conf_matrix

metrics_dict = {
    "Accuracy": accuracy_score(y_test, y_predicted),
    "Precision": precision_score(y_test, y_predicted),
    "Recall": recall_score(y_test, y_predicted),
    "F1 Score": f1_score(y_test, y_predicted),
}

metrics_df = pd.DataFrame(
    metrics_dict.values(), index=metrics_dict.keys(), columns=["Value"]
)

metrics_df

## **Model & Prediction Export**

os.chdir(MODELS_DIR)

kaggle_df = pd.DataFrame(y_predicted, index=X_test.index, columns=["Survived"])
kaggle_df.index.name = "PassengerId"
kaggle_df.replace({False: 0, True: 1}, inplace=True)
kaggle_df.to_csv("kaggle_submission.csv")
kaggle_df.head()
